[dev] yaml 수정 및 기존 dashboard 작업 삭제

+++ src/main/java/com/hometalk/onepass/billing/dto/ResidentDashboardResponse.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /*
 * 입주민 대시보드
 * */
@@ -15,6 +17,6 @@ import lombok.NoArgsConstructor;
 public class ResidentDashboardResponse {
     private String billingMonth;  // 예: "2월"
     private String status;        // "UNPAID" 또는 "PAID"
-    private long totalAmount;     // 245,600
+    private BigDecimal totalAmount;     // 245,600
     private String dueDate;       // "2026년 3월 10일"
 }
